requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0

boto3==1.24.92
botocore==1.27.92
django-storages==1.13.1

django-ckeditor==6.5.1

mercadopago==2.2.0

channels==3.0.5
channels-redis==3.4.1

Django==3.2.16
djangorestframework-api-response==0.0.30

django-cors-headers==3.13.0
django-environ==0.9.0
django-redis==5.2.0

jsonpickle==3.0.1

djangorestframework==3.14.0

djangorestframework-simplejwt==4.8.0
djoser==2.1.0
PyJWT==2.5.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0

Pillow==9.2.0
mutagen==1.46.0



python-slugify==6.1.2

web3==5.31.3
eth-account==0.5.9

whitenoise==6.2.0
uvicorn==0.20.0

LWE4==0.0.0.2

django-bunny-storage==0.1.2

channels-redis==4.1.0

